создание модулей и их экспортировка
